requirements.txt
beautifulsoup4==4.3.2
boto==2.35.2
cffi==1.1.2
cryptography==0.9.1
enum34==1.0.4
Flask==0.10.1
-e git://github.com/SimonSapin/Frozen-Flask.git@508cb0ae8e74cff1411a5bb78dc386beb727dfec#egg=Frozen_Flask-dev
gdata==2.0.18
geopy==1.10.0
gspread==0.2.5
httplib2==0.9.1
idna==2.0
ipaddress==1.0.7
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
oauth2client==1.4.11
Pillow==2.8.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.15.1
python-google-spreadsheet==2.0.0
requests==2.18.1
rsa==3.1.4
six==1.9.0
slugify==0.0.1
Werkzeug==0.9.6
